Add Dockerfile for deploying the application
This commit adds a Dockerfile for deploying the application. The Dockerfile sets up a Ubuntu 22.04 environment with necessary packages such as tightvncserver, xfce4, qtcreator, and vim-tiny. It also creates directories and copies application files. The Dockerfile_entrypoint.sh script is included to configure the VNC server and startup script. The container is started with the entrypoint.sh script.